fix(examples): correct import path for DashJS after moving files
